+++ app/models/maxim.rb
@@ -1,0 +1,7 @@
+class Maxim < ApplicationRecord
+  belongs_to :user
+
+  validates :body, presence: true
+
+  scope :recent, -> { order(created_at: :desc) }
+end